Fix another source of race condition
At first I thought of doing the bare minimum when
`queries_mutex` was locked. However, scheduling
the task does not lock `exec_mutex` but other ones.

So, there is no deadlock in sight of scheduling
within `queries_mutex`. This way, we prevent a rare
case in which the task is scheduled before the query
is even added to the queries list.